[FIX] web: avoid interactions between drag and drop tests
The commit 6b0b998e1a625fdca8c07e11468505f8221c3870 introduces a new option
to drag element without dropping them.

But it's not allowed to start a new drag and drop until the first one is
completely done.

So when the element is removed from the dom (the test is finished), we complete
the drag and drop in order to let next tests perform drag and drop.